refactor(test): rewrite test class.
