+++ packages/core/jest.config.js
@@ -3,5 +3,5 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   roots: ['test/'],
-  // testMatch: ['**/test/**/*.[jt]s?(x)', '**/?(*.)+(spec).[jt]s?(x)'],
+  testMatch: ['**/test/**/*.[jt]s?(x)', '**/?(*.)+(spec).[jt]s?(x)'],
 };
